refactor: extract ErrorCategory/ErrorCode taxonomy to core
Adds src/synthorg/core/error_taxonomy.py owning ErrorCategory, ErrorCode,
CATEGORY_TITLES, CODE_CATEGORY_PREFIX, NOT_FOUND_BAND, category_title(),
and category_type_uri(). Repoints engine/, tools/, budget/, communication/,
integrations/, ontology/, providers/, observability/prometheus_labels.py
and api/dto.py to import the taxonomy from core instead of api.errors.
api/errors.py now imports the taxonomy from core for its concrete
subclasses (those move to core in the next commit).

Renames _CODE_CATEGORY_PREFIX/_NOT_FOUND_BAND to drop the underscore
since they are now cross-module package-public constants.

Adds tests/unit/core/test_error_taxonomy.py covering enum prefix
invariants, four-digit codes, helper coverage, and prefix-table keys.

Refs #1610
