+++ src/synthorg/integrations/errors.py
@@ -12,7 +12,7 @@ should propagate.
 
 from typing import ClassVar
 
-from synthorg.api.errors import ErrorCategory, ErrorCode
+from synthorg.core.error_taxonomy import ErrorCategory, ErrorCode
 
 
 class IntegrationError(Exception):
